INNOVATION OF CARROT CHIPS AS AN ALTERNATIVE TO LESSEN THE INTAKE OF JUNK FOODS OF SELECTED ELEMENTARY STUDENTS
Abstract

This research aims to measure the viability of Carrot Chips as a healthier alternative to junk food for elementary students in order to reduce consumption to unhealthy snacks. In order to lessen the unhealthy effects of junk food, Carrot Chips are an innovation that seeks to provide more nutritional value than junk food to people, especially students. This research selected 15 4th to 6th grade students as the respondents. This research used a quantitative research design in which surveys were given to the respondents after consuming Carrot Chips for a given period of time and comparing its effects to the consumption of junk food. Results showed that most of the respondents found Carrot Chips to be appetizing, energizing, and beneficial for eye health. On the other hand, junk food was found to have negative effects, being associated with low energy and sickliness. In conclusion, Carrot Chips are an effective and healthier alternative to junk food.