FOREIGN OR LOCAL: FACTORS AFFECTING THE CHOICE OF STUDENTS IN BUYING CHOCOLATES
Abstract

Consumers' perception towards international and local brands is based on what they hear, see and experience. This quantitative research was conducted to identify the factors affecting the choice of students in buying chocolates. The data were collected through the use of survey questionnaire with questions constructed by the researchers. This study has 162 respondents from selected Grade 12 students in San Pedro National High School. Descriptive approach was used as the research design of this study. The findings revealed that the main factor affecting the choice of students in buying chocolates is the quality. Moreover, the respondents usually prefer foreign brand of chocolates than the locally-made products.