Bautista, Ashley Mae S. Dela Cruz, Darline Vyan B. Lintag, Precious Wilma C. Reyes, Germaine G. Ylanan, Alyson Miguel C.
Abstract
A successful business has been associated with work ethics. It is a set of values based
on the ideals of discipline and hard work. This qualitative research tackles the influence of
work ethics to the envisioned success of a business. The employees of a well-known milktea
shop in Poblacion Hagonoy, Bulacan served as the participants of the study. The participants
are selected through convenience sampling. Data were collected with the means of face-to-face
interview and by the use of validated interview guide questions which contains 16 items. The
researchers collected data from the participants to share their beliefs, opinions and perceptions
that served as concrete foundation of the study. The study concluded that the work ethics of
employee has an influence to the success of the business. Thus, the statements given by the
participants proved that work ethics is one of the reasons why business is deemed successful.